Add channel type checks

+++ commands/stop.js
@@ -1,7 +1,7 @@
 const { SlashCommandBuilder } = require('discord.js');
 const { useQueue } = require("discord-player");
 
-const { checkInVoice, checkPlaying } = require('../helpers/check');
+const { checkChannelType, checkInVoice, checkPlaying } = require('../helpers/check');
 const { editReply, handleError } = require('../helpers/message');
 
 
@@ -13,9 +13,10 @@ module.exports = {
         await interaction.deferReply();
 
         try {
+            if (!checkChannelType(interaction, client)) return;
             if (!checkInVoice(interaction, client)) return;
             if (!checkPlaying(interaction, client)) return;
-            
+
             const queue = useQueue(interaction.guild.id);
             queue.delete();
